Handle exising .emacs.d dir

+++ setup.sh
@@ -5,5 +5,9 @@ cp bash_aliases ~/.bash_aliases
 
 cp profile ~/.profile
 
-cp -r emacs.d ~/.emacs.d
+if [ -e ~/.emacs.d ]; then
+    cp -fr emacs.d/* ~/.emacs.d
+else
+    cp -r emacs.d ~/.emacs.d    
+fi
